+++ scripts/docs.js
@@ -3,13 +3,14 @@ const jsdoc2md = require('jsdoc-to-markdown');
 const fs = require('fs');
 const path = require('path');
 
-const inputFile = 'src/test.js';
+const inputFile = 'src/**/*.js';
 const outputDir = __dirname;
 
 const BASE_URL = '/Docs/API/';
 //const OUTPUT_PATH_PREFIX = path.resolve(outputDir, './../docs/') + '/';
 const OUTPUT_PATH_PREFIX = 'd:\\MassiveHeights.Blacksmith\\blacksmith2d.io\\Blacksmith.Homepage\\App_Data\\docs-2\\src\\API\\';
 var events = [];
+var globalModuleListText = '';
 
 String.prototype.replaceAll = function (search, replacement) {
   var target = this;
@@ -28,7 +29,9 @@ class Generator {
     let extendsString = this.renderExtends(item);
 
     text += `<h1><span style="color: #999">${kind}</span> ${item.identifier.name} ${extendsString}</h1>\n`;
-    text += this.parseCodeTag(item.identifier.description) + '\n';
+    //text += this.parseCodeTag(item.identifier.description) + '\n';
+
+    text += this.getDesc(item.identifier, false);
 
     if (item.identifier.kind === 'class') {
       text += this.renderMethods(item, 'member');
@@ -43,11 +46,11 @@ class Generator {
     let className = item.name;
     let cat = item.cat;
 
-    let filepath = `${OUTPUT_PATH_PREFIX}${className}.md`;
+    let filepath = `${OUTPUT_PATH_PREFIX}${className}.html`;
 
     if (cat) {
       cat = cat.replace(/\./g, '/');
-      filepath = `${OUTPUT_PATH_PREFIX}${cat}/${className}.md`;
+      filepath = `${OUTPUT_PATH_PREFIX}${cat}/${className}.html`;
     }
 
     ensureDirectoryExistence(filepath);
@@ -109,6 +112,43 @@ class Generator {
     return text;
   }
 
+  renderGlobalModuleList() {
+    let cats = {};
+
+    this.data.forEach(x => {
+      if (x.ignore)
+        return;
+
+      if (x.kind !== 'class' && x.kind !== 'enum')
+        return;
+
+      let cat = this.getTypeCat(x);
+      if (!cat)
+        return;
+
+      if (!(cat in cats))
+        cats[cat] = [];
+
+      cats[cat].push(x);
+    });
+
+    globalModuleListText = '<h2>API Documentation</h2>'
+
+    for (var p in cats) {
+      let cat = p.replace(/\./g, '/');
+      let fnames = cat.split('/');
+      if (fnames.length > 1)
+        continue;
+
+      let url = `${BASE_URL}${cat}`;
+      globalModuleListText += `<a href="${url}" class="class-list-name">${cat}</a>`;
+    }
+
+    let filepath = path.resolve(outputDir, `${OUTPUT_PATH_PREFIX}/README.md`);
+    ensureDirectoryExistence(filepath);
+    fs.writeFileSync(filepath, globalModuleListText);
+  }
+
   renderModuleList() {
     let cats = {};
 
@@ -135,7 +175,6 @@ class Generator {
       let fnames = cat.split('/');
       let fname = fnames[fnames.length - 1];
 
-
       let text = '';
 
       text += `<h2>${cat}</h2>`;
@@ -148,7 +187,7 @@ class Generator {
         text += '\n';
 
         let url = `${BASE_URL}${cat}/${name}`;
-        text += `<a href="${url}" class="method-name">${name}</a>`;
+        text += `<a href="${url}" class="class-list-name">${name}</a>`;
       }
 
       text += '\n';
@@ -156,6 +195,10 @@ class Generator {
       let filepath = path.resolve(outputDir, `${OUTPUT_PATH_PREFIX}${cat}/README.md`);
       ensureDirectoryExistence(filepath);
       fs.writeFileSync(filepath, text);
+
+      filepath = path.resolve(outputDir, `${OUTPUT_PATH_PREFIX}${cat}/config.json`);
+      ensureDirectoryExistence(filepath);
+      fs.writeFileSync(filepath, '{ "icon": "folder-o" }');
     }
   }
 
@@ -223,26 +266,32 @@ class Generator {
           scopes.push('overridden');
 
         sscope = '<span class="method-static">' + scopes.join(' ') + '</span>';
+        let btn = '';
+
+        if (x.params.length > 0)
+          btn = `<div class="expand"></div>`;
 
         let anchor = x.longname.replace(/#/g, '.');
         if (x.kind == 'constructor') {
           let skind = '<span class="id-kind-method">constructor</span>';
-          text += `<span class="id-name"><a name="${anchor}" href="#${anchor}" class="id-href">${sscope}${name}</a>(${this.genParams(x)}) ${skind}</span><br />\n`;
+          text += `<span class="id-name"><a name="${anchor}" href="#${anchor}" class="id-href">${sscope}${name}</a>(${this.genParams(x)}) ${skind}</span>${btn}<br />\n`;
         }
         else {
           let skind = '<span class="id-kind-method">method</span>';
-          let btn = ``;
-
-          if (x.params.length > 0)
-            btn = `<div class="expand"></div>`;
           text += `<span class="id-name"><a name="${anchor}" href="#${anchor}" class="id-href">${name}</a>(${this.genParams(x)}) ${sscope} ${skind}</span>${btn}<br />\n`;
         }
 
-        text += `${this.getDesc(x)}\n`;
+        //text += `${this.getDesc(x, false)}\n`;
+        text += `${this.getDescType(x, true)}\n`;
 
         if (x.params.length > 0) {
-          text += '<div class="method-details">';
-          text += '<br />\n';
+          let long = this.getDescType(x, false);
+
+          text += '<div class="method-details"><br />\n';
+
+          if (long != '')
+            text += `${long}<br />\n<br />\n`;
+
           text += `${this.renderParamsFull(item, x)}`;
           text += this.renderReturn(item, x);
           text += '\n</div>\n';
@@ -271,8 +320,8 @@ class Generator {
       return '';
 
     let text = '';
-    text = `<h2>Messages</h2>`;
 
+    let c = 0;
     if (item.identifier.fires) {
       item.identifier.fires.forEach(x => {
         const result = events.find(e => e.id === x);
@@ -283,93 +332,29 @@ class Generator {
 
         let name = result.name;
         let anchor = result.longname.replace(/#/g, '.');
+        let btn = '';
+
+        let long = this.getDescType(result, false);
+        if (long != '')
+          btn = `<div class="expand"></div>`;
 
         text += '\n<div class="method-info">\n';
-        text += `<span class="id-name"><a name="${anchor}" href="#${anchor}" class="id-href">${name}</a> ${skind}</span><br />\n`;
-        text += `${this.getDesc(result)}<br />\n`;
+        text += `<span class="id-name"><a name="${anchor}" href="#${anchor}" class="id-href">${name}</a> ${skind}</span>${btn}<br />\n`;
+        text += `${this.getDescType(result, true)}<br />\n`;
+
+        if (long != '') {
+          text += '<div class="method-details"><br />\n';
+          text += `${this.getDescType(result, false)}`;
+          text += '\n</div>\n';
+        }
 
         text += '</div>';
+        c++;
       })
     }
 
-
-    return text;
-  }
-
-  renderParamsFull(classItem, item) {
-    let access = item.access;
-    if (!access)
-      access = 'public';
-
-    if (item.params.length == 0)
-      return '';
-
-    let text = '<span class="span-params">Parameters</span><br />';
-    text += `<div class="params">`;
-
-    for (let i = 0; i < item.params.length; i++) {
-      let element = item.params[i];
-
-      if (!element.type) {
-        text += `${element.name} <br />`;
-        console.log('no type for: ' + element.name);
-        continue;
-      }
-
-
-      let typeName = '';
-      for (let k = 0; k < element.type.names.length; k++) {
-        let type = element.type.names[k];
-
-        typeName += `${type.replace('<', '&lt').replace('>', '&gt')}`;
-        if (k != element.type.names.length - 1)
-          typeName += ' | ';
-      }
-
-      let required = element.optional != null && element.optional == true ? 'false' : 'true';
-      let def = element.defaultvalue != null ? element.defaultvalue : '-';
-
-      text += `<span class="param-name">${element.name} : ${this.renderType(classItem, typeName)}</span>`;
-      let desc = this.getDesc(element);
-      if (desc.length > 0)
-        text += `— ${desc} <br />\n`;
-
-      //text += '</tr>\n';
-    }
-
-    text += `</div>`;
-
-    //text += '</table>\n';
-    return text;
-  }
-
-  renderReturn(classItem, item) {
-    let text = '';
-
-    if (!item.returns)
-      return '';
-
-    if (item.returns.length == 0)
-      return '';
-
-    let str = item.kind === 'member' ? 'Type:' : 'Returns';
-    text += `<span class="doc-returns">${str}</span><br /> <span class="doc-returns-type">`;
-
-    var r = item.returns[0];
-
-    for (var k = 0; k < r.type.names.length; k++) {
-      var element = r.type.names[k];
-
-      text += this.renderType(classItem, element);
-
-      if (k != r.type.names.length - 1)
-        text += ' | ';
-    }
-
-    if (r.description)
-      text += ` — ${r.description}`;
-
-    text += '</span>';
+    if (c > 0)
+      text = `<h2>Messages</h2>` + text;
 
     return text;
   }
@@ -448,10 +433,22 @@ class Generator {
           if (typeString.length > 0)
             typeString = ' : ' + typeString;
         }
+        let long = this.getDescType(x, false);
 
         let anchor = x.longname.replace(/#/g, '.');
-        text += `<span class="id-name"><a name="${anchor}" href="#${anchor}" class="id-href">${name}</a> ${typeString} ${sscope} ${skind}</span><br />\n`;
-        text += `${this.getDesc(x)}<br />\n`;
+        let btn = '';
+
+        if (long != '')
+          btn = `<div class="expand"></div>`;
+
+        text += `<span class="id-name"><a name="${anchor}" href="#${anchor}" class="id-href">${name}</a> ${typeString} ${sscope} ${skind}</span>${btn}<br />\n`;
+        text += `${this.getDescType(x, true)}<br />\n`;
+
+        if (long != '') {
+          text += '<div class="method-details"><br />\n';
+          text += `${this.getDescType(x, false)}`;
+          text += '\n</div>\n';
+        }
 
         text += '\n</div>\n';
         count++;
@@ -464,6 +461,84 @@ class Generator {
       else
         text = `<h2>Inherited Properties</h2>` + text;
     }
+
+    return text;
+  }
+
+  renderParamsFull(classItem, item) {
+    let access = item.access;
+    if (!access)
+      access = 'public';
+
+    if (item.params.length == 0)
+      return '';
+
+    let text = '<span class="span-params">Parameters</span><br />';
+    text += `<div class="params">`;
+
+    for (let i = 0; i < item.params.length; i++) {
+      let element = item.params[i];
+
+      if (!element.type) {
+        text += `${element.name} <br />`;
+        console.log('no type for: ' + element.name);
+        continue;
+      }
+
+
+      let typeName = '';
+      for (let k = 0; k < element.type.names.length; k++) {
+        let type = element.type.names[k];
+
+        typeName += `${type.replace('<', '&lt').replace('>', '&gt')}`;
+        if (k != element.type.names.length - 1)
+          typeName += ' | ';
+      }
+
+      let required = element.optional != null && element.optional == true ? 'false' : 'true';
+      let def = element.defaultvalue != null ? element.defaultvalue : '-';
+
+      text += `<span class="param-name">${element.name} : ${this.renderType(classItem, typeName)}</span>`;
+      let desc = this.getDesc(element);
+      if (desc.length > 0)
+        text += ` — ${desc} <br />\n`;
+
+      //text += '</tr>\n';
+    }
+
+    text += `</div>`;
+
+    //text += '</table>\n';
+    return text;
+  }
+
+  renderReturn(classItem, item) {
+    let text = '';
+
+    if (!item.returns)
+      return '';
+
+    if (item.returns.length == 0)
+      return '';
+
+    let str = item.kind === 'member' ? 'Type:' : 'Returns';
+    text += `<span class="doc-returns">${str}</span><br /> <span class="doc-returns-type">`;
+
+    var r = item.returns[0];
+
+    for (var k = 0; k < r.type.names.length; k++) {
+      var element = r.type.names[k];
+
+      text += this.renderType(classItem, element);
+
+      if (k != r.type.names.length - 1)
+        text += ' | ';
+    }
+
+    if (r.description)
+      text += ` — ${r.description}`;
+
+    text += '</span>';
 
     return text;
   }
@@ -576,7 +651,38 @@ class Generator {
       return '';
     }
 
-    return this.parseCodeTag(this.parseLinks(item.description));
+    let desc = item.description;
+    desc = desc.replaceAll('\r', '<br />');
+
+    return this.parseCodeTag(this.parseLinks(desc));
+  }
+
+  getDescType(item, short) {
+    if (!item || !item.description || item.description == 'undefined' || item.description == 'Description') {
+      console.log('Warning: desciption not found', item.longname);
+
+      return '';
+    }
+
+    let desc = item.description;
+
+    let ix = desc.indexOf('\r\r');
+
+    if (short) {
+      if (ix > 0)
+        desc = desc.substring(0, ix);
+    }
+    else {
+      if (ix > 0)
+        desc = desc.substring(ix + 2);
+      else
+        return '';
+    }
+
+    desc = desc.replaceAll('\r', '<br />');
+    desc = this.parseCodeTag(this.parseLinks(desc));
+
+    return desc;
   }
 
   parseCodeTag(desc) {
@@ -721,3 +827,5 @@ for (const item of classNames) {
   g.renderClass(item);
   g.renderModuleList();
 }
+
+g.renderGlobalModuleList();