+++ src/colliders/Collider.js
@@ -15,16 +15,27 @@ class Collider extends Component {
   constructor() {
     super();
 
-    /** @public @type {boolean} Dirty flag */
+    /**
+     * Dirty flag.
+     * @private @type {boolean}
+     */
     this.mChanged = true;
 
-    /** @public @type {Vector} Global in stage coordinates center */
+    /**
+     * Global in stage coordinates center.
+     * @private @type {Vector}
+     */
     this.mCenter = new Vector();
 
-    /** @public @type {Vector} Global in stage coordinates min x and y vertex */
+    /**
+     * Global in stage coordinates min x and y vertex.
+     * @private @type {Vector}
+     */
     this.mMin = new Vector();
 
-    /** @public @type {Vector} Global in stage coordinates max x and y vertex */
+    /**
+     * Global in stage coordinates max x and y vertex. 
+     * @private @type {Vector} */
     this.mMax = new Vector();
   }
 
@@ -47,7 +58,5 @@ class Collider extends Component {
    * @param {Matrix} transform Game object world transformation with zero position.
    * @param {Vector} position  Rigid body position.
    */
-  refresh(transform, position) {
-
-  }
+  refresh(transform, position) { }
 }